add: more flexibility for plasmons in QuotasCalculator
Added an argument for setting the plasmon parameters in the __init__() method of the QuotasCalculator, as well as the possibility of adjusting the parameters by calling the set_up_plasmon_probabilities method. Also actually included the bulk plasmon excitations in the calculation of the yield, which I somehow forgot before.