fix: DielTensor; add: as_dict and from_dict methods for QuotasCalculator
There was an issue with the as_dict and from_dict methods of DielTensor, which became clear when implementing the same methods for the QuotasCalculator class. I fixed these and added a simple version for the QuotasCalculator. This can be improved in the future to include the escape function, or we can simply optimize the set_up_escape_function method.