change: definition of the surface plasmon parameter
The way the surface plasmon parameter is defined currently means that the lower the parameter, the higher the rate of plasmon excitations. This is a little counter intuitive, and inverse to the bulk plasmon parameter logic. I've redefined the surface parameter in such a way that this issue is fixed. Setting the parameter to zero now disables surface plasmon excitations, and increasing it increases the likelihood of these excitations happening.