add: max energy range for energies
The range of energies we now consider is frankly ridiculous, and slows down the calculation quite a bit. I've added a default energy range relative to the Fermi level, based on the fact that the ionization energy is never above 25 eV. Hence, electrons more than 25 eV below the fermi level, or empty states more than 25 eV above the fermi level do not matter, hence they are not used by the QuotasCalculator.